4-3-16
--In the Web.Config file in the root directory, in the <appSettings>
block, the “value” string for the “SendGridPassword” key must be filled
in.

--The code is for the online version. To run on the desktop, do the
following.
(1) In the Web.Config file in the root directory, in the
<connectionStrings> block, comment out the “HussainsDB” entry.
(2) In the “IdentityModels.cs” file in the “Models” folder, comment out
the “HussainsDB” base (with “public ApplicationDbContext” above) and
uncomment the “DefaultConnection” base (with “public
ApplicationDbContext” above).